Minor fixes

+++ vulkan/shaders/quad.frag
@@ -1,7 +1,7 @@
 #version 450
 
 layout(location = 0) out vec4 out_color;
-layout(location = 1) out uvec4 out_uv;
+layout(location = 1) out vec4 out_uv;
 
 layout(push_constant) uniform UBO {
 	layout(offset = 96) vec4 color;
@@ -9,5 +9,5 @@ layout(push_constant) uniform UBO {
 
 void main() {
 	out_color = data.color;
-	out_uv = uvec4(0, 0, 0, 0);
+	out_uv = vec4(0, 0, 0, 0);
 }
